[Artemii] Slight refactor
