+++ lib/constants.dart
@@ -1,1 +1,6 @@
 const String baseUrl = 'https://yummly2.p.rapidapi.com';
+const Map<String, dynamic> headers = {
+  'X-RapidAPI-Host': 'yummly2.p.rapidapi.com',
+  'X-RapidAPI-Key':
+  'ac69a3c921msh8cf410c6cf97884p181dc6jsnd0285b2bcd4c'
+};
