+++ lib/food_page.dart
@@ -5,21 +5,16 @@ import 'package:food/constants.dart';
 import 'package:food/food_cubit.dart';
 import 'package:food/food_repo.dart';
 
-
-void main() {
-  runApp(const MyApp());
-}
-
-class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
+class FoodPage extends StatefulWidget {
+  const FoodPage({Key? key}) : super(key: key);
 
   @override
-  _MyAppState createState() {
-    return _MyAppState();
+  _FoodPageState createState() {
+    return _FoodPageState();
   }
 }
 
-class _MyAppState extends State<MyApp> {
+class _FoodPageState extends State<FoodPage> {
   late FoodCubit _foodCubit;
 
   @override
@@ -32,11 +27,7 @@ class _MyAppState extends State<MyApp> {
             Dio(
               BaseOptions(
                 baseUrl: baseUrl,
-                headers: {
-                  'X-RapidAPI-Host': 'yummly2.p.rapidapi.com',
-                  'X-RapidAPI-Key':
-                      'ac69a3c921msh8cf410c6cf97884p181dc6jsnd0285b2bcd4c'
-                },
+                headers: headers,
               ),
             ),
           ),
@@ -47,156 +38,148 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: 'Food',
-      theme: ThemeData(
-        primaryColor: Colors.white,
-      ),
-      home: BlocBuilder(
-        bloc: _foodCubit,
-        builder: (context, FoodState state) {
-          return Scaffold(
-            appBar: AppBar(
-              centerTitle: true,
-              backgroundColor: Colors.transparent,
-              elevation: 0,
-              title: Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: const [
-                  Icon(Icons.restaurant_menu, color: Colors.black),
-                  SizedBox(width: 10),
-                  Text(
-                    'Food Recipe',
-                    style: TextStyle(
-                      fontSize: 20,
-                      color: Colors.black,
-                      letterSpacing: 1,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
-                ],
+    return Scaffold(
+      appBar: AppBar(
+        centerTitle: true,
+        backgroundColor: Colors.transparent,
+        elevation: 0,
+        title: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: const [
+            Icon(Icons.restaurant_menu, color: Colors.black),
+            SizedBox(width: 10),
+            Text(
+              'Food Recipe',
+              style: TextStyle(
+                fontSize: 20,
+                color: Colors.black,
+                letterSpacing: 1,
+                fontWeight: FontWeight.w500,
               ),
             ),
-            body: state.isLoading
-                ? const Center(child: CircularProgressIndicator())
-                : ListView.builder(
-                    itemCount: state.foods.length,
-                    itemBuilder: (context, index) {
-                      return Container(
-                        margin: const EdgeInsets.symmetric(
-                            horizontal: 22, vertical: 10),
-                        width: MediaQuery.of(context).size.width,
-                        height: 180,
-                        decoration: BoxDecoration(
-                          color: Colors.black,
-                          borderRadius: BorderRadius.circular(15),
-                          boxShadow: [
-                            BoxShadow(
-                              color: Colors.black.withOpacity(0.6),
-                              offset: const Offset(
-                                0.0,
-                                10.0,
-                              ),
-                              blurRadius: 10.0,
-                              spreadRadius: -6.0,
+          ],
+        ),
+      ),
+      body: BlocBuilder(
+        bloc: _foodCubit,
+        builder: (context, FoodState state) {
+          return state.isLoading
+              ? const Center(child: CircularProgressIndicator())
+              : ListView.builder(
+                  itemCount: state.foods.length,
+                  itemBuilder: (context, index) {
+                    return Container(
+                      margin: const EdgeInsets.symmetric(
+                          horizontal: 22, vertical: 10),
+                      width: MediaQuery.of(context).size.width,
+                      height: 180,
+                      decoration: BoxDecoration(
+                        color: Colors.black,
+                        borderRadius: BorderRadius.circular(15),
+                        boxShadow: [
+                          BoxShadow(
+                            color: Colors.black.withOpacity(0.6),
+                            offset: const Offset(
+                              0.0,
+                              10.0,
                             ),
-                          ],
-                          image: DecorationImage(
-                            colorFilter: ColorFilter.mode(
-                              Colors.black.withOpacity(0.2),
-                              BlendMode.multiply,
-                            ),
-                            image: NetworkImage(state.foods[index].images),
-                            fit: BoxFit.cover,
+                            blurRadius: 10.0,
+                            spreadRadius: -6.0,
                           ),
+                        ],
+                        image: DecorationImage(
+                          colorFilter: ColorFilter.mode(
+                            Colors.black.withOpacity(0.2),
+                            BlendMode.multiply,
+                          ),
+                          image: NetworkImage(state.foods[index].images),
+                          fit: BoxFit.cover,
                         ),
-                        child: Stack(
-                          children: [
-                            Align(
-                              child: Padding(
-                                padding:
-                                    const EdgeInsets.symmetric(horizontal: 5.0),
-                                child: Text(
-                                  state.foods[index].name,
-                                  style: const TextStyle(
-                                    color: Colors.white,
-                                    fontSize: 25.0,
-                                    fontWeight: FontWeight.w400,
-                                    letterSpacing: 1,
-                                  ),
-                                  overflow: TextOverflow.ellipsis,
-                                  maxLines: 2,
-                                  textAlign: TextAlign.center,
+                      ),
+                      child: Stack(
+                        children: [
+                          Align(
+                            child: Padding(
+                              padding:
+                                  const EdgeInsets.symmetric(horizontal: 5.0),
+                              child: Text(
+                                state.foods[index].name,
+                                style: const TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 25.0,
+                                  fontWeight: FontWeight.w400,
+                                  letterSpacing: 1,
                                 ),
+                                overflow: TextOverflow.ellipsis,
+                                maxLines: 2,
+                                textAlign: TextAlign.center,
                               ),
-                              alignment: Alignment.center,
                             ),
-                            Align(
-                              child: Row(
-                                mainAxisAlignment:
-                                    MainAxisAlignment.spaceBetween,
-                                children: [
-                                  Container(
-                                    padding: const EdgeInsets.all(5),
-                                    margin: const EdgeInsets.all(10),
-                                    decoration: BoxDecoration(
-                                      color: Colors.black.withOpacity(0.4),
-                                      borderRadius: BorderRadius.circular(15),
-                                    ),
-                                    child: Row(
-                                      children: [
-                                        const Icon(
-                                          Icons.star,
-                                          color: Colors.yellow,
-                                          size: 18,
-                                        ),
-                                        const SizedBox(width: 7),
-                                        Text(
-                                          state.foods[index].rating.toString(),
-                                          style: const TextStyle(
-                                            color: Colors.white,
-                                            fontSize: 14,
-                                          ),
-                                        ),
-                                      ],
-                                    ),
+                            alignment: Alignment.center,
+                          ),
+                          Align(
+                            child: Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                Container(
+                                  padding: const EdgeInsets.all(5),
+                                  margin: const EdgeInsets.all(10),
+                                  decoration: BoxDecoration(
+                                    color: Colors.black.withOpacity(0.4),
+                                    borderRadius: BorderRadius.circular(15),
                                   ),
-                                  Container(
-                                    padding: const EdgeInsets.all(5),
-                                    margin: const EdgeInsets.all(10),
-                                    decoration: BoxDecoration(
-                                      color: Colors.black.withOpacity(0.4),
-                                      borderRadius: BorderRadius.circular(15),
-                                    ),
-                                    child: Row(
-                                      children: [
-                                        const Icon(
-                                          Icons.schedule,
-                                          color: Colors.yellow,
-                                          size: 18,
+                                  child: Row(
+                                    children: [
+                                      const Icon(
+                                        Icons.star,
+                                        color: Colors.yellow,
+                                        size: 18,
+                                      ),
+                                      const SizedBox(width: 7),
+                                      Text(
+                                        state.foods[index].rating.toString(),
+                                        style: const TextStyle(
+                                          color: Colors.white,
+                                          fontSize: 14,
                                         ),
-                                        const SizedBox(width: 7),
-                                        Text(
-                                          state.foods[index].totalTime,
-                                          style: const TextStyle(
-                                            color: Colors.white,
-                                            fontSize: 14,
-                                          ),
-                                        ),
-                                      ],
-                                    ),
+                                      ),
+                                    ],
                                   ),
-                                ],
-                              ),
-                              alignment: Alignment.bottomLeft,
+                                ),
+                                Container(
+                                  padding: const EdgeInsets.all(5),
+                                  margin: const EdgeInsets.all(10),
+                                  decoration: BoxDecoration(
+                                    color: Colors.black.withOpacity(0.4),
+                                    borderRadius: BorderRadius.circular(15),
+                                  ),
+                                  child: Row(
+                                    children: [
+                                      const Icon(
+                                        Icons.schedule,
+                                        color: Colors.yellow,
+                                        size: 18,
+                                      ),
+                                      const SizedBox(width: 7),
+                                      Text(
+                                        state.foods[index].totalTime,
+                                        style: const TextStyle(
+                                          color: Colors.white,
+                                          fontSize: 14,
+                                        ),
+                                      ),
+                                    ],
+                                  ),
+                                ),
+                              ],
                             ),
-                          ],
-                        ),
-                      );
-                    },
-                  ),
-          );
+                            alignment: Alignment.bottomLeft,
+                          ),
+                        ],
+                      ),
+                    );
+                  },
+                );
         },
       ),
     );
