+++ lib/main.dart
@@ -1,0 +1,31 @@
+
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:food/food_page.dart';
+
+void main() {
+  WidgetsFlutterBinding.ensureInitialized();
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO create file with routes
+    SystemChrome.setPreferredOrientations([
+      DeviceOrientation.portraitUp,
+      DeviceOrientation.portraitDown,
+    ]);
+    return MaterialApp(
+      locale: const Locale('az'),
+      title: 'Food',
+      initialRoute: '/',
+      debugShowCheckedModeBanner: false,
+      routes: {
+        '/': (context) => const FoodPage(),
+      },
+    );
+  }
+}
